+++ CHANGES.md
@@ -1,5 +1,6 @@
 [unreleased]
 * update documentation
+* now includes all type of error, exception, shutdown handling
 
 #### 0.5.0 / 2022-08-02
 * converted original Gist to repo